+++ aten/src/ATen/CMakeLists.txt
@@ -98,7 +98,6 @@ file(GLOB miopen_h CONFIGURE_DEPENDS "miopen/*.h")
 file(GLOB miopen_cpp CONFIGURE_DEPENDS "miopen/*.cpp")
 
 file(GLOB mkl_cpp CONFIGURE_DEPENDS "mkl/*.cpp")
-file(GLOB mklrng_cpp CONFIGURE_DEPENDS "mklrng/*.cpp")
 file(GLOB mkldnn_cpp CONFIGURE_DEPENDS "mkldnn/*.cpp")
 
 file(GLOB mkldnn_xpu_h CONFIGURE_DEPENDS "native/mkldnn/xpu/*.h" "native/mkldnn/xpu/detail/*.h")
@@ -523,7 +522,6 @@ if(USE_LIGHTWEIGHT_DISPATCH)
 endif()
 if(AT_MKL_ENABLED)
   set(all_cpu_cpp ${all_cpu_cpp} ${mkl_cpp})
-  set(all_cpu_cpp ${all_cpu_cpp} ${mklrng_cpp})
 endif()
 if(AT_KLEIDIAI_ENABLED)
   set(all_cpu_cpp ${all_cpu_cpp} ${native_kleidiai})
